refactor(annotation-server): moving tests
